src/requirements.txt
fastapi
uvicorn
pydantic
passlib
pyjwt
tortoise-orm
aiosqlite